backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.9.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
PyMuPDF==1.25.5
openpyxl==3.1.5
lxml==5.4.0
python-docx==1.1.2
aiofiles==24.1.0
